analysis-model/requirements.txt
pandas
langdetect
spacy
sentence-transformers
scikit-learn
google-generativeai
